Rename newgraph.js to graph.js
